refactor: move packages under pkg to internals
use github.com/fatih/color to make some text be bold
