+++ cmd/list.go
@@ -3,7 +3,7 @@ package cmd
 import (
 	"fmt"
 
-	"github.com/musaubrian/tinygo/pkg/model"
+	"github.com/musaubrian/tinygo/internal/model"
 	"github.com/spf13/cobra"
 )
