+++ internal/utils/dir_test.go
@@ -1,0 +1,13 @@
+package utils_test
+
+import (
+	"testing"
+
+	"github.com/musaubrian/tinygo/internal/utils"
+)
+
+func TestCreateDir(t *testing.T) {
+	if utils.CreateDir() != nil {
+		t.Error("Expected nil got an error")
+	}
+}
