--- pkg/utils/dir_test.go
@@ -1,9 +1,0 @@
-package utils
-
-import "testing"
-
-func TestCreateDir(t *testing.T) {
-	if CreateDir() != nil {
-		t.Error("Expected nil got an error")
-	}
-}
